requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
Pillow>=9.0.0

# Для вебхука на Vercel (FastAPI + serverless)
fastapi>=0.112.0
uvicorn>=0.30.0


